requirements.txt
yfinance>=0.2.40
pandas-ta>=0.3.14b
pandas>=2.0
pyyaml>=6.0
rich>=13.0
python-dotenv==1.0.1
matplotlib>=3.7